2.6 수업내용

+++ Ch09_Extends/src/ch09/abstract11/PartTimeEmployee.java
@@ -1,10 +1,11 @@
 package ch09.abstract11;
+
 // 일용직 사원
 public class PartTimeEmployee extends Employee {
 	
-	protected int dailyPay;		// 일당
-	protected int workDay;		// 일한 일수
-
+	protected int dailyPay;	// 일당
+	protected int workDay;	// 일한 일수
+	
 	public PartTimeEmployee(String empno, String name,
 			int dailyPay, int workDay) {
 		super(empno, name);
@@ -14,17 +15,18 @@ public class PartTimeEmployee extends Employee {
 
 	@Override
 	public double getMonthPay() {
-		double mPay = dailyPay * workDay;
+		double mPay = (double)(dailyPay*workDay);
 		return mPay;
 	}
-
+	
 	@Override
 	public void showEmployeeInfo() {
 		super.showEmployeeInfo();
 		System.out.println("dailyPay : " + dailyPay);
 		System.out.println("workDay : " + workDay);
-		System.out.printf("monthPay : %.2f\n ", getMonthPay());
+		System.out.printf("monthPay : %.2f\n " , getMonthPay());
 	}
-	
+
+
 
 }
